# api/openapi/swagger.yaml
components:
  schemas:
    errs.Error:
      properties:
        code:
          $ref: '#/components/schemas/errs.ErrorCode'
        message:
          type: string
        params:
          items:
            $ref: '#/components/schemas/errs.Param'
          type: array
          uniqueItems: false
      type: object
    errs.ErrorCode:
      type: integer
      x-enum-varnames:
      - ErrorCodeOK
      - ErrorCodeCanceled
      - ErrorCodeUnknown
      - ErrorCodeInvalidArgument
      - ErrorCodeDeadlineExceeded
      - ErrorCodeNotFound
      - ErrorCodeAlreadyExists
      - ErrorCodePermissionDenied
      - ErrorCodeResourceExhausted
      - ErrorCodeFailedPrecondition
      - ErrorCodeAborted
      - ErrorCodeOutOfRange
      - ErrorCodeUnimplemented
      - ErrorCodeInternal
      - ErrorCodeUnavailable
      - ErrorCodeDataLoss
      - ErrorCodeUnauthenticated
      - ErrorCodeClosedRequest
    errs.Param:
      properties:
        key:
          type: string
        value:
          type: string
      type: object
...
    handlers.PostCreateDTO:
      properties:
        body:
          type: string
      type: object
    handlers.PostDTO:
      properties:
        body:
          type: string
        created_at:
          type: string
        id:
          type: string
        updated_at:
          type: string
      type: object
    handlers.PostListDTO:
      properties:
        count:
          type: integer
        items:
          items:
            $ref: '#/components/schemas/handlers.PostDTO'
          type: array
          uniqueItems: false
      type: object
    handlers.PostUpdateDTO:
      properties:
        body:
          type: string
        id:
          type: string
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  title: example
  version: 0.0.0
openapi: 3.1.0
paths:
...
  /api/v1/posts/posts/:
    get:
      parameters:
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: page_number
        schema:
          type: integer
      - in: query
        name: order_by
        schema:
          items:
            type: string
          type: array
          uniqueItems: false
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.PostListDTO'
          description: Filtered list of posts
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: List of posts
      tags:
      - post
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.PostCreateDTO'
        description: Create post request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.PostDTO'
          description: Created post
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Create post
      tags:
      - post
  /api/v1/posts/posts/{id}:
    delete:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Delete post by id
      tags:
      - post
    get:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.PostDTO'
          description: Requested post
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get post by id
      tags:
      - post
    patch:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.PostUpdateDTO'
        description: Update post request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.PostDTO'
          description: Updated post
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Update post
      tags:
      - post
...
servers:
- url: http://127.0.0.1:8000/